Remove 2 unused extension methods to BlocClosure
In the case of #valueWithoutUserInterruption, it was not used.

In the case of #valueWithoutTermination, it uses #ensure: to
value the block and return the result if there was no exception,
and ignoring the exception and then returning nil. I think it can
be replaced by just using #valueNoContextSwitch as #ensure:
uses in the case the primitive fails.